The methods for finding offset between frames were separated
